+++ src/main/scala/millfork/env/Pointy.scala
@@ -6,14 +6,23 @@ trait Pointy {
   def name: Option[String]
   def indexType: VariableType
   def elementType: VariableType
+  def readOnly: Boolean
 }
 
 case class StackVariablePointy(offset: Int, indexType: VariableType, elementType: VariableType) extends Pointy {
   override def name: Option[String] = None
+  override def readOnly: Boolean = false
 }
 
 case class VariablePointy(addr: Constant, indexType: VariableType, elementType: VariableType, zeropage: Boolean) extends Pointy {
   override def name: Option[String] = None
+  override def readOnly: Boolean = false
 }
 
-case class ConstantPointy(value: Constant, name: Option[String], size: Option[Int], indexType: VariableType, elementType: VariableType, alignment: MemoryAlignment) extends Pointy
+case class ConstantPointy(value: Constant,
+                          name: Option[String],
+                          size: Option[Int],
+                          indexType: VariableType,
+                          elementType: VariableType,
+                          alignment: MemoryAlignment,
+                          override val readOnly: Boolean) extends Pointy